Add utils module, with distance computation
The newly added utils.py module allows to compute distances from a query set to a dataset
using either brute force queries or a given FAISS index. The distances are then manipulated
so to adhere to the ones stored in the ann-benchmarks HDF5 files.

This funcionality has been integrated in the load_data module. In particular, if only part
of a dataset is loaded, then baseline distances are recomputed on the fly, because the
distances stored in the HDF5 file are relative to the _entire_ dataset.