address parallelization issues in metrics.get_epsilons
As outlined in issue #11 Python has issues with multithreading
on large datasets.
This commit tries to address this by delegating the parallel processing
to the C++ code of FAISS. In fact, if we pass an index instead of a dataset
to the function utils.compute_distances then the distance computation is
delegated to the index. In particular, using a "flat" index runs all
the queries in parallel.

We can then use Numpy's vectorized operations to combine all the results.